+++ src/main/groovy/com/rapidminer/gradle/PluginConfiguration.groovy
@@ -24,6 +24,5 @@ package com.rapidminer.gradle;
 public class PluginConfiguration {
 	String id
 	String description
-	String displayName
 	def tags = []
 }