+++ src/main/groovy/com/rapidminer/gradle/RapidMinerGradlePlugin.groovy
@@ -131,9 +131,6 @@ class RapidMinerGradlePlugin implements Plugin<Project> {
 				if(!extension.description){
 					throw new RuntimeException('Missing extension description!')
 				}
-				if(!extension.displayName){
-					throw new RuntimeException('Missing display name!')
-				}
 
 				pluginBundle {
 					website = 'https://www.rapidminer.com'
@@ -147,7 +144,6 @@ class RapidMinerGradlePlugin implements Plugin<Project> {
 					plugins {
 						rapidminerPlugin {
 							id = "org.rapidminer.${extension.id}"
-							displayName = extension.displayName
 						}
 					}
 				}