+++ CHANGES.md
@@ -1,7 +1,7 @@
 ## Change Log
 
 #### 0.3.0
-* Changes plugin release mechanism from Bintray to new 'com.gradle.plugin-publish' plugin
+* Changes plugin publishing mechanism from Bintray to new 'com.gradle.plugin-publish' plugin
 * Removes deprecated plugin properties files
 
 #### 0.2.10